Remove nonsense annotations from IEnumVariant.
It doesn't make sense to put [Out] on an IntPtr without the ref
or out keyword - Mono's COM code can't do anything with that, and
it asserts. Presumably Framework ignores this, but since it's an
internal class and there's no reason to expect that a real
application does this, it seems reasonable to just fix it here.